# perf(faster): Upgrade Rspack to 1.3.6+, fix persistent cache perf issue