## Motivation

The Rspack persistent cache was enabled in Docusaurus with https://github.com/facebook/docusaurus/pull/10931

Unfortunately, the performance improvement wasn't significant because code-generated files would always recompile/reload, even if their content/hash didn't change. 

After our feedback in https://github.com/web-infra-dev/rspack/issues/9413, with v1.3.6, Rspack is changing its default strategy for the persistent cache, using a file content hash instead of the filesystem lastModified time.

This PR significantly improves the efficiency of the Rspack persistent cache, so that warm rebuilds will bundle the app faster, by avoiding recompiling all the code-generated files.

Many code-generated files are generated by Docusaurus core in `website/.docusaurus`, but they can also be generated by plugins that generate MDX files (often the case of docsgen plugin like OpenAPI).

This upgrade affects the speed of both the dev server and the production build.

On our own website, we use a custom [changelog plugin](https://docusaurus.io/changelog/) which is a fork of the blog plugin that does MDX codegen. We can see that it's now faster on prod builds, and the persistent cache also significantly improve the startup time of the dev server (on 2nd startups).


## Benchmarks

Impact of the persistent cache on total build times:
- Small site: 15.8s -> 9.3s (previously 10.4s)
- Medium site: 37.4s -> 24.9s (previously 29.7s)

Note: 
- Those numbers represent the impact on the average build time, and not just the bundling phase
- The bundling phase is likely to be at least x2-x3 faster with the Rspack persistent cache enabled
- Other Docusaurus phases (such as the SSG phase, which happens just after bundling) are also expensive to run.
- The persistent cache of Rspack is still in active development and will keep improving in the future (see https://rspack.dev/blog/announcing-1-2#persistent-cache)

### Cold build

Small site

```bash
yarn clear:website
hyperfine --runs 5 'DOCUSAURUS_NO_PERSISTENT_CACHE=true yarn build:website:fast'

Benchmark 1: DOCUSAURUS_NO_PERSISTENT_CACHE=true yarn build:website:fast
  Time (mean ± σ):     15.878 s ±  1.077 s    [User: 49.951 s, System: 8.975 s]
  Range (min … max):   14.507 s … 17.140 s    5 runs
```

Medium site

```bash
yarn clear:website
hyperfine --runs 5 'DOCUSAURUS_NO_PERSISTENT_CACHE=true yarn build:website --locale en'

Benchmark 1: DOCUSAURUS_NO_PERSISTENT_CACHE=true yarn build:website --locale en
  Time (mean ± σ):     37.432 s ±  1.609 s    [User: 130.666 s, System: 20.958 s]
  Range (min … max):   34.742 s … 38.823 s    5 runs
```

### Warm build - before upgrading

Small site

```bash
yarn clear:website
yarn build:website:fast
hyperfine --runs 5 'yarn build:website:fast'

Benchmark 1: yarn build:website:fast
  Time (mean ± σ):     10.442 s ±  0.878 s    [User: 33.773 s, System: 7.245 s]
  Range (min … max):    9.524 s … 11.599 s    5 runs
```

Medium site

```bash
yarn clear:website
yarn build:website --locale en
hyperfine --runs 5 'yarn build:website --locale en'

Benchmark 1: yarn build:website --locale en
  Time (mean ± σ):     29.721 s ±  2.407 s    [User: 103.784 s, System: 19.332 s]
  Range (min … max):   27.340 s … 33.297 s    5 runs
```


### Warm build - after upgrading

Small site

```bash
yarn clear:website
yarn build:website:fast
hyperfine --runs 5 'yarn build:website:fast'

Benchmark 1: yarn build:website:fast
  Time (mean ± σ):      9.309 s ±  1.063 s    [User: 30.241 s, System: 7.210 s]
  Range (min … max):    7.975 s … 10.314 s    5 runs
```

Medium site


```bash
yarn clear:website
yarn build:website --locale en
hyperfine --runs 5 'yarn build:website --locale en'

Benchmark 1: yarn build:website --locale en
  Time (mean ± σ):     24.992 s ±  3.595 s    [User: 96.350 s, System: 19.542 s]
  Range (min … max):   22.522 s … 31.272 s    5 runs
```

## Test Plan

Local benchmarks on my M3 computer:


It should also speed up CI rebuilds

### Test links

https://deploy-preview-11114--docusaurus-2.netlify.app/