@slorber 2025-04-18
Will merge it later, once v1.3.6 is officially released